backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.9
requests==2.31.0
pydantic-settings==2.2.1
